+++ src/TestingApp/Controllers/QuizControllers/QuizController.cs
@@ -23,7 +23,7 @@ public class QuizController : ControllerBase
 
 
     [HttpPut("{id}")]
-    public async ValueTask<IActionResult> UpdateAsync(int id, QuizForCreationDTO quizForUpdateDTO)
+    public async ValueTask<IActionResult> UpdateAsync(long id, QuizForCreationDTO quizForUpdateDTO)
            => Ok(await quizService.UpdateAsync(id, quizForUpdateDTO));
 
 
@@ -33,11 +33,11 @@ public class QuizController : ControllerBase
 
 
     [HttpGet("{id}")]
-    public async ValueTask<IActionResult> GetAsync([FromRoute] int id)
+    public async ValueTask<IActionResult> GetAsync([FromRoute] long id)
           => Ok(await quizService.GetAsync(u => u.Id == id));
     
     
     [HttpDelete("{id}")]
-    public async Task<IActionResult> DeleteAsync([FromRoute] int id)
+    public async Task<IActionResult> DeleteAsync([FromRoute] long id)
            => Ok(await quizService.DeleteAsync(id));
 }
